fixed return tag on divide function

+++ src/main/java/com/codingblackfemales/Calculator.java
@@ -13,7 +13,7 @@ public class Calculator {
         return firstInput * secondInput;
     }
 
-    public static Integer divide(final int firstInput, final int secondInput) {
-        return firstInput / secondInput;
+    public static Double divide(final int firstInput, final int secondInput) {
+        return (double)firstInput / secondInput;
     }
 }
